--- a/packages/bitcore-node/src/models/transaction.ts
+++ b/packages/bitcore-node/src/models/transaction.ts
@@ -15,7 +15,7 @@ import { EventStorage } from './events';
 import { Libs } from '../providers/libs';
 import { BaseTransaction, ITransaction } from './baseTransaction';
 import { Readable, Transform } from 'stream';
-import { Collection, BulkWriteOpResultObject } from 'mongodb';
+import { Collection } from 'mongodb';
 import { partition } from '../utils/partition';
 
 export { ITransaction };
@@ -166,23 +166,15 @@ export class MempoolTxEventTransform extends Transform {
 }
 
 export class MongoWriteStream extends Transform {
-  promiseChain = (Promise.resolve() as unknown) as Promise<BulkWriteOpResultObject[]>;
   constructor(private collection: Collection) {
     super({ objectMode: true });
   }
 
   async _transform(data: Array<any>, _, done) {
-    this.promiseChain = this.promiseChain.then(() =>
-      Promise.all(
-        partition([...data], data.length / Config.get().maxPoolSize).map(batch => this.collection.bulkWrite(batch))
-      )
+    await Promise.all(
+      partition([...data], data.length / Config.get().maxPoolSize).map(batch => this.collection.bulkWrite(batch))
     );
     done(null, data);
-  }
-
-  async _flush(done) {
-    await this.promiseChain;
-    done();
   }
 }