--- a/packages/bitcore-node/test/helpers/index.ts
+++ b/packages/bitcore-node/test/helpers/index.ts
@@ -11,15 +11,17 @@ import { RateLimitStorage } from '../../src/models/rateLimit';
 import { EventStorage } from '../../src/models/events';
 
 export async function resetDatabase() {
-  console.log('Restting database');
-  await resetModel(BitcoinBlockStorage);
-  await resetModel(TransactionStorage);
-  await resetModel(CoinStorage);
-  await resetModel(WalletAddressStorage);
-  await resetModel(WalletStorage);
-  await resetModel(StateStorage);
-  await resetModel(RateLimitStorage);
-  await resetModel(EventStorage);
+  console.log('Resetting database');
+  await Promise.all([
+    resetModel(BitcoinBlockStorage),
+    resetModel(TransactionStorage),
+    resetModel(CoinStorage),
+    resetModel(WalletAddressStorage),
+    resetModel(WalletStorage),
+    resetModel(StateStorage),
+    resetModel(RateLimitStorage),
+    resetModel(EventStorage)
+  ]);
 }
 
 export async function resetModel(model: BaseModel<any>) {