I'll see if we can address some of the speed issues from #2459 

Closes #2459

Currently trying: 
* Using partition and mapping batches to bulkWrite

# This Branch
```
................................................................................
Benchmark for 80 (1 MB) blocks completed after 44.073 s
1.8151702856624237 MB/s
0.5509125 Seconds/Block

.....
Benchmark for 5 (32 MB) blocks completed after 79.835 s
2.0041335253961297 MB/s
15.966999999999999 Seconds/Block

.
Benchmark for 1 (64 MB) blocks completed after 21.152 s
3.02571860816944 MB/s
21.152 Seconds/Block
```

# Master 
```
................................................................................
Benchmark for 80 (1 MB) blocks completed after 95.397 s
0.8386007945742529 MB/s
1.1924625 Seconds/Block

.....
Benchmark for 5 (32 MB) blocks completed after 159.296 s
1.004419445560466 MB/s
31.859199999999998 Seconds/Block
```

# 324944c
```
................................................................................
Benchmark for 80 (1 MB) blocks completed after 42.783 s
1.869901596428488 MB/s
0.5347875 Seconds/Block

.....
Benchmark for 5 (32 MB) blocks completed after 81.814 s
1.9556555112816878 MB/s
16.3628 Seconds/Block
```